+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Fixed a rounding issue causing the world boundary to be offset by 1 block
+
 ## 0.0.20
 *2024-08-08 19:52*
 - Added proper block texture rendering